Split tests into S3FileReferenceStoreS3SpecificIT and S3PartitionStoreIT
